Refactoring: Move method
